requirements.txt
openai
replicate
python-dotenv
sounddevice
scipy
streamlit
mistralai
pillow